Fix tanggal kas report, fix respon kas report print, reduce duration available when promo room removed

+++ app/src/main/java/livs/code/frontoffice/data/remote/respons/SalesItemListResponse.kt
@@ -29,7 +29,7 @@ data class SaleItemList(
 	val jumlah: Int = 0,
 
 	@field:SerializedName("Total")
-	val total: Int = 0,
+	val total: Long = 0,
 
 	@field:SerializedName("nama_item")
 	val namaItem: String = "",
